Mw/add cart back end

+++ Controllers/ShoppingCartController.php
@@ -35,8 +35,7 @@ class ShoppingCartController extends AbstractController
     //This method adds a product to the shopping cart
     public function addProductToShoppingCart(){
 
-        //überarbeiten
-        $id = $_REQUEST['id'];
+        $id = $this->request->GET('id');
 
         if($this->shoppingCart->checkProductInShoppingCart($id)){
             $this->shoppingCart->increaseAmountOfProduct($id);
@@ -61,15 +60,21 @@ class ShoppingCartController extends AbstractController
         $this->request->setSESSION("shoppingCart", serialize($this->shoppingCart));
 
         header("Location: http://Localhost/WebundMultimedia/product/$id");
+        die();
 
     }
 
     //This method deletes a product from the shopping cart
     public function deleteProductFromShoppingCart(){
-        $id = $_REQUEST['id'];
+        $id = $this->request->GET('id');
+        $delete = $this->request->GET('delete');
 
         if($this->shoppingCart != ""){
-            if($this->shoppingCart->checkProductInShoppingCart($id)){
+            if($this->shoppingCart->checkProductInShoppingCart($id) && $delete == "n"){
+                $this->shoppingCart->decreaseAmountOfProduct($id);
+                $this->request->setSESSION("shoppingCart", serialize($this->shoppingCart));
+            }
+            elseif ($this->shoppingCart->checkProductInShoppingCart($id) && $delete == "y"){
                 $this->shoppingCart->deleteProduct($id);
                 $this->request->setSESSION("shoppingCart", serialize($this->shoppingCart));
             }
@@ -78,6 +83,7 @@ class ShoppingCartController extends AbstractController
         header("Location: http://Localhost/WebundMultimedia/shoppingcart");
         die();
     }
+
 
     //This method empties the shopping cart
     public function emptyShoppingCart(){
